Move CameraController.h to ship/
